+++ shuffle.h
@@ -1,6 +1,9 @@
 #ifndef SHUFFLE_H
 #define SHUFFLE_H
 
-void shuffle(double *, int);
+#include <stdlib.h> //rand
+#include <stddef.h>
+
+void shuffle(double *, size_t);
 
 #endif /* SHUFFLE_H */
